Move agents listprovider
