+++ modules/admin-service/src/main/java/org/opencastproject/adminui/endpoint/CaptureAgentsEndpoint.java
@@ -33,9 +33,9 @@ import org.opencastproject.adminui.util.TextFilter;
 import org.opencastproject.capture.CaptureParameters;
 import org.opencastproject.capture.admin.api.Agent;
 import org.opencastproject.capture.admin.api.AgentState;
+import org.opencastproject.capture.admin.api.AgentsListQuery;
 import org.opencastproject.capture.admin.api.CaptureAgentStateService;
 import org.opencastproject.index.service.util.RestUtils;
-import org.opencastproject.list.common.query.AgentsListQuery;
 import org.opencastproject.security.api.SecurityService;
 import org.opencastproject.security.api.UnauthorizedException;
 import org.opencastproject.security.util.SecurityUtil;
