+++ modules/admin-service/src/main/java/org/opencastproject/adminui/endpoint/ListProvidersEndpoint.java
@@ -26,6 +26,7 @@ import static org.opencastproject.adminui.endpoint.EndpointUtil.generateJSONObje
 
 import org.opencastproject.adminui.exception.JsonCreationException;
 import org.opencastproject.authorization.xacml.manager.impl.list.AclsListQuery;
+import org.opencastproject.capture.admin.api.AgentsListQuery;
 import org.opencastproject.index.service.util.JSONUtils;
 import org.opencastproject.index.service.util.RestUtils;
 import org.opencastproject.list.api.DefaultResourceListQuery;
@@ -33,7 +34,6 @@ import org.opencastproject.list.api.ListProviderException;
 import org.opencastproject.list.api.ListProviderNotFoundException;
 import org.opencastproject.list.api.ListProvidersService;
 import org.opencastproject.list.api.ResourceListQuery;
-import org.opencastproject.list.common.query.AgentsListQuery;
 import org.opencastproject.list.common.query.EventListQuery;
 import org.opencastproject.list.common.query.GroupsListQuery;
 import org.opencastproject.list.common.query.JobsListQuery;
