+++ modules/capture-admin-service-api/src/main/java/org/opencastproject/capture/admin/api/AgentsListProvider.java
@@ -1,0 +1,28 @@
+/*
+ * Licensed to The Apereo Foundation under one or more contributor license
+ * agreements. See the NOTICE file distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ *
+ * The Apereo Foundation licenses this file to you under the Educational
+ * Community License, Version 2.0 (the "License"); you may not use this file
+ * except in compliance with the License. You may obtain a copy of the License
+ * at:
+ *
+ *   http://opensource.org/licenses/ecl2.txt
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ *
+ */
+
+package org.opencastproject.capture.admin.api;
+
+public interface AgentsListProvider {
+  String PROVIDER_PREFIX = "AGENTS";
+  String NAME = AgentsListProvider.PROVIDER_PREFIX + ".NAME";
+  String STATUS = AgentsListProvider.PROVIDER_PREFIX + ".STATUS";
+}
